Delete the discovery/ package, move contents to servicediscovery.mdns
