+++ babble/src/main/java/io/mosaicnetworks/babble/fragments/join/WebRTCJoinGroupFragment.java
@@ -45,9 +45,9 @@ import androidx.annotation.NonNull;
 import io.mosaicnetworks.babble.R;
 import io.mosaicnetworks.babble.fragments.BaseConfigActivity;
 import io.mosaicnetworks.babble.fragments.OnFragmentInteractionListener;
-import io.mosaicnetworks.babble.discovery.HttpPeerDiscoveryRequest;
-import io.mosaicnetworks.babble.discovery.Peer;
-import io.mosaicnetworks.babble.discovery.ResponseListener;
+import io.mosaicnetworks.babble.servicediscovery.mdns.HttpPeerDiscoveryRequest;
+import io.mosaicnetworks.babble.node.Peer;
+import io.mosaicnetworks.babble.servicediscovery.mdns.ResponseListener;
 import io.mosaicnetworks.babble.node.ConfigManager;
 import io.mosaicnetworks.babble.node.GroupDescriptor;
 import io.mosaicnetworks.babble.service.BabbleService;
