+++ babble/src/main/java/io/mosaicnetworks/babble/node/BabbleNode.java
@@ -28,10 +28,9 @@ import android.util.Log;
 
 import com.google.gson.JsonSyntaxException;
 
-import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 
-import io.mosaicnetworks.babble.discovery.PeersProvider;
+import io.mosaicnetworks.babble.servicediscovery.mdns.PeersProvider;
 import mobile.Mobile;
 import mobile.Node;
