+++ babble/src/main/java/io/mosaicnetworks/babble/node/Peer.java
@@ -22,9 +22,47 @@
  * SOFTWARE.
  */
 
+package io.mosaicnetworks.babble.node;
+
+import com.google.gson.annotations.SerializedName;
+
 /**
- * io.mosaicnetworks.babble.discovery contains some discovery support classes. The
- * HttpPeerDiscovery classes implement a server for publishing information about a node. The
- * Peers classes define Peers records required to start a Babble instance.
+ * An immutable class representing a Babble peer.
  */
-package io.mosaicnetworks.babble.discovery;
+public final class Peer {
+
+    /**
+     * The peer's public key
+     */
+    @SerializedName("PubKeyHex")
+    public final String pubKeyHex;
+
+    /**
+     * The peer's network address
+     */
+    @SerializedName("NetAddr")
+    public final String netAddr;
+
+    /**
+     * The peer's moniker
+     */
+    @SerializedName("Moniker")
+    public final String moniker;
+
+    /**
+     * Constructor
+     * @param pubKeyHex public key as created by the KeyPair class
+     * @param netAddr the network address on which the peer can be contacted
+     * @param moniker a moniker, this does not need to be unique across all peer's in a network
+     */
+    public Peer(String pubKeyHex, String netAddr, String moniker) {
+
+        if (pubKeyHex==null || netAddr==null || moniker==null) {
+            throw new NullPointerException("Null arguments are not accepted");
+        }
+
+        this.pubKeyHex = pubKeyHex;
+        this.netAddr = netAddr;
+        this.moniker = moniker;
+    }
+}
