+++ babble/src/main/java/io/mosaicnetworks/babble/servicediscovery/mdns/PeersProvider.java
@@ -22,47 +22,22 @@
  * SOFTWARE.
  */
 
-package io.mosaicnetworks.babble.discovery;
-
-import com.google.gson.annotations.SerializedName;
+package io.mosaicnetworks.babble.servicediscovery.mdns;
 
 /**
- * An immutable class representing a Babble peer.
+ * A provider of peer information
  */
-public final class Peer {
+public interface PeersProvider {
 
     /**
-     * The peer's public key
+     * Provide genesis peers
+     * @return a json formatted string of the genesis peers
      */
-    @SerializedName("PubKeyHex")
-    public final String pubKeyHex;
+    String getGenesisPeers();
 
     /**
-     * The peer's network address
+     * Provide current peers
+     * @return a json formatted string of current peers
      */
-    @SerializedName("NetAddr")
-    public final String netAddr;
-
-    /**
-     * The peer's moniker
-     */
-    @SerializedName("Moniker")
-    public final String moniker;
-
-    /**
-     * Constructor
-     * @param pubKeyHex public key as created by the KeyPair class
-     * @param netAddr the network address on which the peer can be contacted
-     * @param moniker a moniker, this does not need to be unique across all peer's in a network
-     */
-    public Peer(String pubKeyHex, String netAddr, String moniker) {
-
-        if (pubKeyHex==null || netAddr==null || moniker==null) {
-            throw new NullPointerException("Null arguments are not accepted");
-        }
-
-        this.pubKeyHex = pubKeyHex;
-        this.netAddr = netAddr;
-        this.moniker = moniker;
-    }
+    String getCurrentPeers();
 }
